+++ src/components/Card/Card.stories.tsx
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react';
 import Card from './Card';
 import Typography from '../Typography/Typography';
 
-storiesOf('Interface|Card', module).add('basic', () => (
+storiesOf('Atoms|Card', module).add('basic', () => (
   <Card>
     <div style={{ padding: 40 }}>
       <Typography>Card body</Typography>