+++ src/components/Checkbox/Checkbox.stories.tsx
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import Checkbox from './Checkbox';
 
-storiesOf('Forms|Checkbox', module).add('basic', () => (
+storiesOf('Atoms|Checkbox', module).add('basic', () => (
   <Checkbox defaultChecked={false} onChange={action('change')}>
     'I agree to the Privacy Policy and Terms and Conditions'
   </Checkbox>