+++ src/components/Icon/Icon.stories.tsx
@@ -2,4 +2,6 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import Icon from './Icon';
 
-storiesOf('Interface|Icon', module).add('warn', () => <Icon name="warn" />);
+storiesOf('Atoms|Icons', module)
+  .add('warn', () => <Icon name="warn" />)
+  .add('close', () => <Icon name="close" />);