+++ src/components/Input/Input.stories.tsx
@@ -1,6 +1,7 @@
 import React, { FormEvent, useState } from 'react';
 import { storiesOf } from '@storybook/react';
 import Input from './Input';
+import FieldAction from '../FieldAction/FieldAction';
 
 const InputStory = () => {
   const [value, setValue] = useState('');
@@ -15,7 +16,7 @@ const InputStory = () => {
   );
 };
 
-storiesOf('Forms|Input', module)
+storiesOf('Atoms|Input', module)
   .add('basic', () => {
     return <InputStory />;
   })
@@ -24,13 +25,21 @@ storiesOf('Forms|Input', module)
       <Input label="Form field" value="" disabled onChange={() => false} />
     );
   })
-  .add('error', () => {
+  .add('with action', () => {
     return (
-      <Input
-        label="Form field"
-        value="Filled Text Here"
-        error="error"
-        onChange={() => false}
-      />
+      <div style={{ width: 400, marginTop: 100 }}>
+        <Input
+          label="Form field"
+          value="Filled Text Here"
+          error
+          onChange={() => false}
+          postfix={() => (
+            <FieldAction color="pink" icon="warn">
+              Password must be more than 8 characters and contain both numbers
+              and letters
+            </FieldAction>
+          )}
+        />
+      </div>
     );
   });