+++ src/components/Modal/Modal.tsx
@@ -2,15 +2,15 @@ import React from 'react';
 import css from './Modal.module.scss';
 import { ReactNode } from 'react';
 import Icon from '../Icon/Icon';
-import Typography from '../Typography/Typography';
 import Overlay from '../Overlay/Overlay';
 
 export interface ModalProps {
-  header?: ReactNode;
+  header?: () => ReactNode;
   children: ReactNode;
   close: () => void;
   isOpen: boolean;
   shouldOverlayClose?: boolean;
+  hideCloseButton?: boolean;
 }
 
 const Modal = ({
@@ -19,6 +19,7 @@ const Modal = ({
   header,
   children,
   shouldOverlayClose = true,
+  hideCloseButton,
 }: ModalProps) => {
   const handleClickOverlay = () => {
     if (shouldOverlayClose) {
@@ -28,14 +29,12 @@ const Modal = ({
   return (
     <Overlay active={isOpen} onClick={handleClickOverlay}>
       <div className={css.modal}>
-        <button className={css.close} type="button" onClick={close}>
-          <Icon name="close" width={20} height={20} />
-        </button>
-        {header && (
-          <div className={css.header}>
-            <Typography type="subtitle">{header}</Typography>
-          </div>
+        {!hideCloseButton && (
+          <button className={css.close} type="button" onClick={close}>
+            <Icon name="close" width={20} height={20} />
+          </button>
         )}
+        <div className={css.header}>{header && header()}</div>
         <div className={css.body}>{children}</div>
       </div>
     </Overlay>