+++ src/components/ToggleContent/index.ts
@@ -1,1 +1,0 @@
-export { default as ToggleContent } from './ToggleContent';