+++ CHANGELOG.md
@@ -1,5 +1,59 @@
 # Changelog
 
+## [0.16.0]
+### Updated
+- Modal - change header from string to render function
+## [0.15.0]
+### Added
+- Field Action
+- Color palette
+### Updated
+- Input - add postfix
+
+## [0.14.0]
+### Updated
+- Modal - add classname property
+
+## [0.13.1]
+### Fixed
+- Modal - empty header style
+
+## [0.13.0]
+### Added
+- Buttons - add secondary ghost
+
+## [0.12.0]
+### Updated
+- Typography - add classname property
+
+## [0.11.0]
+### Updated
+- Typography - add custom color 
+### Fixed
+- Input - fix autofill font 
+- Modal - fix header color
+
+## [0.10.0]
+### Updated
+- Input - add error message
+
+## [0.9.1]
+#### Fixed
+- Checkbox - return event instead value, pass base input props
+
+## [0.9.0]
+### Updated
+- Rename stories for Atomic Design
+### Fixed
+- Checkbox - replace defaultChecked for the checked
+
+## [0.8.0]
+### Fixed
+- Add typings to package
+- Remove stories file from declaration
+### Removed
+- React styleguidist
+
 ## [0.7.0]
 ### Updated
 - Button - add ghost theme